fix: do not leak existing username/email on regular sign in form
The HomeIdpDiscoveryAuthenticator sets the user in the auth context
in order to function properly. However, it doesn't clear it in case
it "fails", meaning the next authenticator will be aware of the user.

This is problematic when a username/email doesn't match an idp, and
keycloak redirects to the username password form. Indeed, the latter
behaves differently whether the user is known or not (showing/hiding
the email field). Without this clear, we basically leak information
on which user exists.

This fix should ensure the username/password form always shows the
email field.