fix: Do not prevent restart of authentication flow when showing IdP selection dialog
The first login attempt had set the login_hint parameter in the session.
When restarting the flow the login_hint is still present in the session.
If bypassing of the login page is enabeld, no option was given to change
the username, but the login_hint from the previous attempt was reused.

Solution is to no longer store the username as a login_hint. Need to
check if this breaks something else, but automated tests look good so
far.

Closes #292